feat: Add theme preference functionality
- Added a new import statement for the `MyMauiTemplate.Utilities` namespace.
- Modified the `App` class constructor to set the `UserAppTheme` property using the `PreferencesHelper.GetTheme()` method.

This commit adds functionality to allow users to set their preferred theme in the app. The `PreferencesHelper.GetTheme()` method retrieves the user's selected theme from storage, and it is then assigned to the `UserAppTheme` property in the constructor of the `App` class. This ensures that the app will display with the user's chosen theme upon launch.