Update Projects Structure, Cleaned up Some UI, Added SQLite DB support
Updated projects structure to be separated by what they contain and handle, Also added SQLite DB support that works on Windows UWP but not on android so currently not using it in the project but the stub is there for when it can be implemented correctly for all platforms.  Lots of other changes and clean up.